# Conflicts:
#	BrotatoPrototype/Assets/Prefabs/Abilities/AbilityHpRegen.prefab
#	BrotatoPrototype/Assets/Prefabs/Abilities/AbilityMaxHp.prefab
#	BrotatoPrototype/Assets/Prefabs/Enemies/Charger.prefab
#	BrotatoPrototype/Assets/Prefabs/Enemies/Enemy_Shooter.prefab
#	BrotatoPrototype/Assets/Prefabs/Enemies/Enemy_Stump.prefab
#	BrotatoPrototype/Assets/Prefabs/Enemies/ProjectileEnemy.prefab
#	BrotatoPrototype/Assets/Prefabs/Weapons/Bow.prefab
#	BrotatoPrototype/Assets/Prefabs/Weapons/Sownya.prefab
#	BrotatoPrototype/Assets/Prefabs/Weapons/Sword.prefab
#	BrotatoPrototype/Assets/TestForTeam/Mark/Weap_Bow_anim.fbx.meta
#	BrotatoPrototype/Assets/TestForTeam/Mark/Weap_Bow_attack.fbx.meta
#	BrotatoPrototype/Assets/TestForTeam/Mark/Weapon Anim/Weap_Bow_anim.fbx
#	BrotatoPrototype/Assets/TestForTeam/Mark/Weapon Anim/Weap_Bow_anim.fbx.meta

+++ BrotatoPrototype/Assets/Scripts/Enemy/EnemyHealth.cs
@@ -41,11 +41,11 @@ public class EnemyHealth : LivingEntity
         base.TakeHit(damage, isCrit);
         if (isCrit)
         {
-            TemporaryMessageManager.Instance.AddMessageOnScreen(damage.ToString() + "!", this.gameObject.transform.position, Color.yellow, 0.5f, 20);
+            TemporaryMessageManager.Instance.AddMessageOnScreen(((int)damage).ToString() + "!", this.gameObject.transform.position, Color.yellow, 0.5f, 20);
         }
         else
         {
-            TemporaryMessageManager.Instance.AddMessageOnScreen(damage.ToString(), this.gameObject.transform.position, Color.white, 0.5f, 20);
+            TemporaryMessageManager.Instance.AddMessageOnScreen(((int)damage).ToString(), this.gameObject.transform.position, Color.white, 0.5f, 20);
         }
         
     }
